requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
langchain-groq>=0.1.6 # Added for Groq integration
langchain-core>=0.2.14
langgraph>=0.0.60
python-dotenv>=1.0.1
email_validator>=2.1.1
pydantic-settings>=2.3.4
pydantic>=2.7.4
pytest>=8.4.1